+++ src/Application/Application.csproj
@@ -18,7 +18,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Engine\Engine.csproj" />
-    <ProjectReference Include="..\Messaging\Messaging.csproj" />
     <ProjectReference Include="..\Features\MetadataProcessor\MetadataProcessor.csproj" />
   </ItemGroup>
 </Project>
