+++ src/Features/MetadataProcessor/Events/MetadataProcessedEvents.cs
@@ -1,0 +1,15 @@
+namespace Kurmann.Videoschnitt.Features.MetadataProcessor.Events;
+
+public class MetadataProcessedEvent
+{
+    public MetadataProcessedEvent(string message)
+    {
+        Message = message;
+    }
+
+    public string Message { get; }
+
+    public List<FileInfo> ProcessedFiles { get; } = new();
+}
+
+public class ProcessMetadataRequest {}
