+++ src/Features/MetadataProcessor/Handler/MetadataProcessedEventHandler.cs
@@ -1,3 +1,4 @@
+using Kurmann.Videoschnitt.Features.MetadataProcessor.Events;
 using Wolverine;
 
 namespace Kurmann.Videoschnitt.Features.MetadataProcessor.Handler;
@@ -9,4 +10,10 @@ public class MetadataProcessedEventHandler
         Console.WriteLine($"Metadata processed: {message.Message}");
         // Weitere Logik zur Verarbeitung der Nachricht
     }
+
+    public void Handle(ProcessMetadataRequest request)
+    {
+        Console.WriteLine("Processing metadata...");
+        // Weitere Logik zur Verarbeitung der Nachricht
+    } 
 }
